backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
langchain==0.1.16
langchain-community==0.0.34
chromadb==0.4.24
sentence-transformers==2.7.0
faiss-cpu==1.7.4
numpy==1.26.4
scikit-learn==1.4.1.post1
requests==2.31.0
httpx==0.27.0
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
motor==3.3.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
structlog==24.1.0
redis==5.0.3
